backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0

# AI & LangChain
google-generativeai>=0.7.0,<0.8.0
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0

# Database
supabase==2.7.0
psycopg2-binary==2.9.9
pgvector==0.3.0

# Google APIs
google-auth==2.34.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.143.0

# Web Push
pywebpush==2.0.0

# Utilities
pillow==10.4.0
python-multipart==0.0.9
aiofiles==24.1.0
httpx==0.27.0
